Fix IE issues when using useFetch with string argument (#5)
The failing internet explorer errors on Object.getOwnProperties (as opposed to returning an array of string characters in Chrome, Edge, and other post proterozoic era web browsers) if a non-object is passed in. When destructuring, the compiled lazy-fetch uses Object.getOwnProperties in the prepareHeaders function. This breaks if a string is passed into use-fetch or use-lazy-fetch.

The solution was to move the string check before the destructuring, which also meant having to manually assign our destructured variables to the prepareHeaders function scope.